refactor(upgrade): extract binary-upgrade modules from daemon/ into src/upgrade/
Moves 6 source files + their mirrored tests into the new src/upgrade/ domain:
- daemon/myco-release-resolver.ts → upgrade/release-resolver.ts
- daemon/apply-binary-update.ts   → upgrade/apply-binary.ts
- daemon/apply-update.ts          → upgrade/orchestrator.ts
- daemon/update-installer.ts      → upgrade/spawn.ts
- daemon/update-in-progress.ts    → upgrade/in-progress.ts
- install/release-assets.ts       → upgrade/release-assets.ts

Intra-domain relative imports flipped; external importers re-pointed to
@myco/upgrade/*. Exports spawnApplyUpgrade (renamed from private
spawnApplyUpdate) and resolveOrchestratorBinary added to spawn.ts.
The __apply-update CLI dispatch string is preserved unchanged.
