+++ tests/daemon/update-checker-github.test.ts
@@ -65,7 +65,7 @@ import {
   setDevBuildCliEntry,
   type CachedCheck,
 } from '@myco/daemon/update-checker.js';
-import { mycoReleasesApiUrl } from '@myco/install/release-assets.js';
+import { mycoReleasesApiUrl } from '@myco/upgrade/release-assets.js';
 import { UPDATE_CONFIG_PATH } from '@myco/constants/update.js';
 
 // ---------------------------------------------------------------------------
